Move XINFO reply decoders into their data types' companion objects
toXStreamInfo/toXGroupInfo/toXConsumerInfo (plus xStreamInfoKnownFields
and the nested toEntry helper) lived as private methods on BaseRedis,
alongside ~150 other command implementations - decoding logic for a
data type belongs with that type's definition instead, matching how
ScriptOutputType already does this (its companion holds its own
Lettuce-conversion logic, see effects.scala). BaseRedis now just calls
XStreamInfo.fromLettuce/XGroupInfo.fromLettuce/XConsumerInfo.fromLettuce.

Also fixed while moving toXStreamInfo: its nested toEntry function
parsed a completely different shape (a single stream entry) from the
flat-field-map logic it was nested inside, unlike its siblings
toXGroupInfo/toXConsumerInfo which are already single-purpose. It's now
its own top-level private function in XStreamInfo's companion, plus a
cleaned-up single `optional` helper replacing the double-Option-wrapped
firstEntry/lastEntry lookups, and dropped a pointless `.view` that
bought no laziness benefit before being immediately forced by `.toMap`.
